+++ CHANGELOG.md
@@ -20,3 +20,4 @@
 - Cache some HTTP requests
 - Remove advanced backup. We're looking into how to update and make this better for the future
 - Change the way that languages are loaded, allowing for easier translations going forward
+- Add in ability to update curse mods from edit mods dialog